added webpack as a module builder
