+++ index.html
@@ -2,7 +2,7 @@
 <html>
  <head>
   <meta charset="UTF-8">
-  <script src="index.js"></script>
+  <script src="bundle.js"></script>
   <title>title</title>
  </head>
  <body>
