+++ package.json
@@ -24,5 +24,8 @@
     "eslint": "^1.5.1",
     "eslint-plugin-react": "^3.5.0",
     "mocha": "^2.3.3"
+  },
+  "dependencies": {
+    "webpack": "^1.12.2"
   }
 }
