+++ src/card.js
@@ -1,0 +1,6 @@
+var Card = {
+    text: "",
+    action: null
+};
+
+module.exports = Card;
