+++ webpack.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    entry: "./src/index.js",
+    output: {
+        path: __dirname,
+        filename: "bundle.js"
+    },
+    module: {
+        loaders: [
+            { test: /\.css$/, loader: "style!css" }
+        ]
+    }
+};
